+++ src/main/scala/farseek/core/FarseekCoreMod.scala
@@ -1,5 +1,6 @@
 package farseek.core
 
+import farseek.core.MethodReplacements.replacements
 import farseek.util.Logging
 import java.io._
 import java.net.URL
@@ -10,6 +11,7 @@ import net.minecraftforge.fml.relauncher.IFMLLoadingPlugin.SortingIndex
 import net.minecraftforge.fml.relauncher._
 import scala.collection.JavaConversions._
 import scala.io.Source._
+import scala.language.reflectiveCalls
 
 /** Core mod class for Farseek that allows method replacements.
   * @see [[farseek.FarseekMod]] for non-core mod class.
@@ -22,9 +24,16 @@ import scala.io.Source._
   *
   * `repose.block.FallingBlockExtensions net.minecraft.block.Block onBlockAdded func_176213_c (Lnet/minecraft/world/World;Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/block/state/IBlockState;)V`
   *
-  * This will redirect all method calls of `net.minecraft.block.Block.onBlockAdded(World, BlockPos, IBlockState)` to `repose.block.FallingBlockExtensions.onBlockAdded(Block, World, BlockPos, IBlockState)`.
-  * Note that the replacement method has the same name and parameters as the original one but is static (on a Scala Object) and has the method owner as an additional initial parameter if the original call was an instance method.
-  * Method calls originating from within the replacement method will NOT be redirected, allowing to call the original behavior as needed.
+  * This will redirect all method calls of `net.minecraft.block.Block.onBlockAdded(World, BlockPos, IBlockState)` to
+  * `repose.block.FallingBlockExtensions.onBlockAdded(Block, World, BlockPos, IBlockState)`.
+  *
+  * Note that the replacement method has the same name and parameters as the original one but is static (on a Scala Object) and
+  * has the method owner as an additional initial parameter if the original method was an instance method.
+  * Method calls originating from within the replacement method will NOT be redirected, which allows calling the original behavior as needed.
+  *
+  * Limitations:
+  *  - Methods must be replaced in the class where they are defined. If you want to change behavior only for a subclass, you still need to replace the top-level method and then match on the first parameter to handle the subtype.
+  *  - No redirection will occur for calls targeting an [[https://docs.oracle.com/javase/tutorial/java/javaOO/anonymousclasses.html anonymous subclass]] of the redirected method.
   *
   * @author delvr
   */
@@ -38,7 +47,9 @@ class FarseekCoreMod extends IFMLLoadingPlugin {
   val getSetupClass = null
   val getModContainerClass = null
   val getAccessTransformerClass = null
-  val getASMTransformerClass = Array(classOf[FarseekClassTransformer].getName)
+  val getASMTransformerClass =
+    if(Package.getPackage("org.spongepowered") != null) Array(classOf[FarseekClassTransformer].getName, classOf[FarseekSpongeClassTransformer].getName)
+    else Array(classOf[FarseekClassTransformer].getName)
 
   def injectData(data: java.util.Map[String, AnyRef]) = {
     gameDir = new File(data("mcLocation").toString)
@@ -48,9 +59,53 @@ class FarseekCoreMod extends IFMLLoadingPlugin {
 
 class FarseekClassTransformer extends IClassTransformer with Logging {
 
+  private var checkedSponge = false
+
+  def transform(obfuscatedName: String, deobfuscatedName: String, bytecode: Array[Byte]): Array[Byte] = {
+    moveAfterSponge()
+    if(bytecode == null) null
+    else new FarseekClassVisitor(bytecode, internalName(deobfuscatedName), replacements).patch
+  }
+
+  /** Sponge mixins can be very strict in their expectations of the code they're patching, and a call redirected by Farseek
+    * will be unrecognized by Sponge and cause a crash if mandatory. Since Farseek is more lenient about patching modified code,
+    * we move it after the Sponge proxy. (This cannot be done with @SortIndex since Sponge places itself last manually.)
+    * We also need to exclude the Farseek transformer from Sponge's pre-mixin transformations performed by the proxy.
+    * For these pre-mixin transformations, we use the FarseekSpongeClassTransformer which only transforms Sponge mixin classes.
+    */
+  private def moveAfterSponge(): Unit = {
+    if(!checkedSponge) {
+      checkedSponge = true
+      val transformers = new java.util.ArrayList(Launch.classLoader.getTransformers)
+      if(transformers.exists(_.getClass.getName == "org.spongepowered.asm.mixin.transformer.Proxy")) {
+        info("Moving Farseek transformer after Sponge proxy")
+        val wrapper = transformers.find(_.toString.contains(classOf[FarseekClassTransformer].getName)).get
+        transformers.remove(wrapper)
+        transformers.add(wrapper)
+        val transformersField = Launch.classLoader.getClass.getDeclaredField("transformers")
+        transformersField.setAccessible(true)
+        transformersField.set(Launch.classLoader, transformers)
+        info("Excluding Farseek transformer from Sponge pre-mixin transformations")
+        val mixinEnvironmentClass = Class.forName("org.spongepowered.asm.mixin.MixinEnvironment")
+        val mixinEnvironment = mixinEnvironmentClass.getDeclaredMethod("getCurrentEnvironment").invoke(null)
+        mixinEnvironmentClass.getDeclaredMethod("addTransformerExclusion", classOf[String]).invoke(mixinEnvironment, classOf[FarseekClassTransformer].getName)
+      }
+    }
+  }
+}
+
+class FarseekSpongeClassTransformer extends IClassTransformer {
+  def transform(obfuscatedName: String, deobfuscatedName: String, bytecode: Array[Byte]): Array[Byte] = {
+    if(bytecode == null || !deobfuscatedName.startsWith("org.spongepowered.common.mixin")) bytecode
+    else new FarseekClassVisitor(bytecode, internalName(deobfuscatedName), replacements).patch
+  }
+}
+
+object MethodReplacements extends Logging {
+
   private val ReplacementsFilepath = "META-INF/farseek_cm.cfg"
 
-  private lazy val replacements: Map[ReplacedMethod, MethodReplacement] = {
+  lazy val replacements: Map[ReplacedMethod, MethodReplacement] = {
     val modsDir = new File(gameDir, "mods")
     val allReplacements = ((files(modsDir) ++ files(new File(modsDir, mcVersion))).flatMap(methodReplacements) ++
       Launch.classLoader.getResources(ReplacementsFilepath).flatMap(methodReplacements)).toSet
@@ -60,7 +115,7 @@ class FarseekClassTransformer extends IClassTransformer with Logging {
   }
 
   private def methodReplacements(file: File): Seq[(ReplacedMethod, MethodReplacement)] = {
-    //trace(s"Checking $file for method replacements")
+    trace(s"Checking $file for method replacements")
     if(!file.getName.endsWith(".jar")) Seq()
     else logged(file, using(new ZipFile(file))(zipFile =>
       Option(zipFile.getEntry(ReplacementsFilepath)).fold(Seq[(ReplacedMethod, MethodReplacement)]())(entry =>
@@ -82,9 +137,6 @@ class FarseekClassTransformer extends IClassTransformer with Logging {
     if(replacements.nonEmpty) debug(s"Found method replacements in $source:\n  ${replacements.mkString("\n  ")}")
     replacements
   }
-
-  def transform(obfuscatedName: String, deobfuscatedName: String, bytecode: Array[Byte]) =
-    if(bytecode == null) null else new FarseekClassVisitor(bytecode, internalName(deobfuscatedName), replacements).patch
 }
 
 case class ReplacedMethod(className: String, methodName: String, descriptor: String)